sy 17:42

+++ 3rd_Project/src/main/java/mall/controller/OrderCompletedController.java
@@ -1,6 +1,11 @@
 package mall.controller;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import javax.servlet.ServletContext;
 import javax.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -10,7 +15,11 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import mall.model.OrderBean;
 import mall.model.OrderDao;
+import mall.model.OrdersProduct;
 import mall.model.TempCart;
+import product.model.ProductBean;
+import users.model.UsersBean;
+import users.model.UsersDao;
 
 
 @Controller
@@ -20,7 +29,10 @@ public class OrderCompletedController {
 	private String viewPage = "orderCompleted";
 	
 	@Autowired
-	private OrderDao dao; 
+	private OrderDao dao;
+	
+	@Autowired
+	private ServletContext application;
 	
 	@RequestMapping(value = command, method = RequestMethod.GET)
 	public String list(
@@ -32,6 +44,8 @@ public class OrderCompletedController {
 			@RequestParam(value = "o_phone2", required = false) String o_phone2,
 			@RequestParam(value = "o_phone3", required = false) String o_phone3,
 			
+			@RequestParam(value = "o_message", required = false) String o_message,
+			
 			@RequestParam(value = "addr1", required = false) String addr1,
 			@RequestParam(value = "addr2", required = false) String addr2,
 			@RequestParam(value = "addr3", required = false) String addr3,
@@ -41,19 +55,35 @@ public class OrderCompletedController {
 			@RequestParam(value = "way", required = false) String way,
 			@RequestParam(value = "pay_bank", required = false) String pay_bank,
 			@RequestParam(value = "pay_name", required = false) String pay_name,
+			@RequestParam(value = "deli", required = false) String deli,
+			@RequestParam(value = "totalPrice", required = false) String totalPrice,
+			@RequestParam(value = "totalPoint", required = false) String totalPoint,
 			
-			HttpServletRequest request,
-			Model model) {
+			HttpServletRequest request, Model model) {
 		 
-		System.out.println("orderCompleted.mall 찾아옴");
 		OrderBean ob = new OrderBean();
+		
+		ob.setU_id(u_id);
+		ob.setO_phone(o_phone1+"-"+o_phone2+"-"+o_phone3);
+		ob.setO_name(o_name);
+		ob.setO_message(o_message);
+		
+		if(addr1.length() < 4) {
+			UsersBean ub = dao.selectUser(u_id);
+			ob.setO_addr(ub.getU_address());
+		} else {
+			addr1 = addr1.replace(",", "");
+			addr2 = addr2.replace(",", "");
+			addr5 = addr5.replace(",", "");
+			ob.setO_addr(addr1+", "+addr2+", "+addr5);
+		}
 		
 		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 		String currentDate = LocalDate.now().format(formatter);
+		ob.setO_date(currentDate);
 		
-		ob.setU_id(u_id);
-		ob.setO_name(o_name);
-		ob.setWay(way);
+		ob.setO_status("주문완료");
+		
 		if(pay_bank.equals("")) {
 			ob.setPay_bank("다른 결제 방법 선택");
 			ob.setPay_name("다른 결제 방법 선택");
@@ -61,30 +91,29 @@ public class OrderCompletedController {
 			ob.setPay_bank(pay_bank);
 			ob.setPay_name(pay_name);
 		}
-	 
-		ob.setO_date(currentDate);
-		ob.setO_phone(o_phone1+"-"+o_phone2+"-"+o_phone3);
-		ob.setO_addr(addr1+addr2+addr3+addr4+addr5);
-		 System.out.println(cart_num+"cart_num 제발");
-		 int MaxO_num = 0;
-		 MaxO_num = dao.selectMaxO_num();
-		 
-		 MaxO_num = MaxO_num+1;
+		
+		ob.setWay(way);
+		
+		int MaxO_num = 0;
+		MaxO_num = dao.selectMaxO_num();
+		MaxO_num = MaxO_num+1;
 		 
 		ob.setO_num(MaxO_num);
 		
-		if(ob.getWay().equals("무통장 입금")) {
+		if(!(Boolean)application.getAttribute("flag")) {
 			TempCart tc =dao.selectTemp(cart_num); 
-			
 			dao.insertOrderProd(tc,MaxO_num);
-			
 			dao.insertOrderInfo(ob);
-		
 			dao.downStockPord(tc);
 			dao.upSalevolumePord(tc);
-		}
+			dao.deleteTemp(cart_num);
+			application.setAttribute("flag", true);
+		} 
 		
-		dao.deleteTemp(cart_num);
+		model.addAttribute("ob", ob); // 주문 정보가 담긴 객체 '주문완료' 페이지로 전달.
+		model.addAttribute("deli", deli); // 배송비 정보 담아서 전달
+		model.addAttribute("totalPrice", totalPrice); // 총 합계금액 정보 담아서 전달
+		model.addAttribute("totalPoint", totalPoint); // 총 포인트 적립금액 정보 담아서 전달
 		
 		return viewPage;
 	}
